requirements.txt
feedparser
langchain
langchain-openai
langchain-community
python-dotenv
bs4
python-docx